+++ src/db/tournaments.ts
@@ -1,13 +1,12 @@
 /// <reference lib="deno.unstable" />
 export const kv = await Deno.openKv();
-import {} from "@/src/apiTypes.ts";
 
 export interface dbTournament {
   slug: string;
 }
 
-export async function createTournament(id: string, slug: string) {
-  const tournamentKey = ["tournaments", id];
+export async function createTournament(slug: string) {
+  const tournamentKey = slug.split("/");
 
   const atomicOp = kv.atomic()
     .check({ key: tournamentKey, versionstamp: null })
@@ -18,5 +17,5 @@ export async function createTournament(id: string, slug: string) {
 }
 
 export function readTournamentList() {
-  return kv.list<dbTournament>({ prefix: ["tournaments"] });
+  return kv.list<dbTournament>({ prefix: ["tournament"] });
 }